feat: 리다이렉트 기능 구현
